% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcc2018pkg.R
\docType{package}
\name{gcc2018pkg}
\alias{gcc2018pkg}
\alias{gcc2018pkg-package}
\title{Infrastructure Tools Package for GCC2018Pkg}
\description{
This package collects various scripts and templates
we've found to be useful in organizing projects.
}
